requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
click==6.7
configparser==3.5.0
constantly==15.1.0
enum34==1.1.6
gdata==2.0.18
google-api-python-client==1.5.5
httplib2==0.9.2
incremental==16.10.1
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauth2client==4.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pylint==1.7.2
python-dotenv==0.6.5
python-gflags==3.1.0
pytz==2016.10
rsa==3.4.2
singledispatch==3.4.0.3
six==1.10.0
twilio==5.6.0
Twisted==16.6.0
uritemplate==3.0.0
wrapt==1.10.11
zope.interface==4.3.3
